:clown_face: identity & mock caenv
Mocks of ABCs were separated into single files and moved to tests directory.
